use ip.go file and optimize the ipv6 ula gen function
